requirements.txt
asgiref==3.5.2
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
defusedxml==0.7.1
Django==4.1
django-allauth==0.40.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
whitenoise==6.2.0
